refactor(LottoPolicy): 전체적으로 사용되는 로또 관련 상수 클래스 LottoPolicy 추가
전체적으로 사용되는 LottoPolicy 클래스를 만들고, domain/policy에 두었습니다.
