+++ src/main/java/lotto/domain/lotto_numbers/BonusNumber.java
@@ -1,8 +1,8 @@
 package lotto.domain.lotto_numbers;
 
 import static lotto.domain.lotto_numbers.LottoNumbersErrorMessages.INVALID_RANGE_OF_LOTTO_NUMBER;
-import static lotto.domain.lotto_numbers.LottoNumbersPolicy.MAX_NUMBER_OF_LOTTO;
-import static lotto.domain.lotto_numbers.LottoNumbersPolicy.MIN_NUMBER_OF_LOTTO;
+import static lotto.domain.policy.LottoPolicy.MAX_NUMBER_OF_LOTTO;
+import static lotto.domain.policy.LottoPolicy.MIN_NUMBER_OF_LOTTO;
 
 public class BonusNumber {
