+++ src/main/java/lotto/domain/lotto_numbers/LottoNumbersErrorMessages.java
@@ -1,6 +1,6 @@
 package lotto.domain.lotto_numbers;
 
-public class LottoNumbersErrorMessages {
+class LottoNumbersErrorMessages {
 
     static final String INVALID_SIZE_OF_LOTTO_NUMBERS = "로또 숫자는 6개만 입력할 수 있습니다.";
     static final String NOT_ALLOW_DUPLICATED_LOTTO_NUMBER = "중복 숫자는 입력될 수 없습니다.";
