--- src/main/java/lotto/domain/lotto_numbers/LottoNumbersPolicy.java
@@ -1,8 +1,0 @@
-package lotto.domain.lotto_numbers;
-
-public class LottoNumbersPolicy {
-
-    static final int MIN_NUMBER_OF_LOTTO = 1;
-    static final int MAX_NUMBER_OF_LOTTO = 45;
-    static final int NUMBER_OF_LOTTO_NUMBERS = 6;
-}
