+++ src/main/java/lotto/domain/policy/LottoPolicy.java
@@ -1,0 +1,10 @@
+package lotto.domain.policy;
+
+public class LottoPolicy {
+
+    public static final int MIN_NUMBER_OF_LOTTO = 1;
+    public static final int MAX_NUMBER_OF_LOTTO = 45;
+    public static final int COUNT_OF_LOTTO_NUMBERS = 6;
+
+    public static final int PRICE_OF_LOTTO = 1_000;
+}
